Allow restoring the Empire of China
The Empire of China can be dismantled, for example if a non-nomadic non-Chinese tag full-annexes Ming. Allow a Chinese or nomadic country to take the decision to claim the Mandate of Heaven even if the Empire of China had been dismantled earlier, assuming they meet all other requirements, such as owning the highlighted provinces.