Let AI seize the Mandate
If AIs are prevented from taking the Mandate in a war, what tends to happen is that even if Ming loses hard, they still remain EoC, but stuck at zero Mandate and perma-bankrupt; ineffectual but taking many decades to kill. This means that after a Mingsplosion the EoC is basically over, as Ming would never recover and even Chinese AIs would always fullannex them, which dismantles the EoC. This makes Asia less flavorful. If the problem with taking the Mandate is that it is not worth 50 warscore, then the right approach is to buff the benefits of taking the Mandate so that it is worth it.