Rebalance Celestial Reforms
Rebalanced modifiers for both flavor and mechanics. Some effects in Vanilla are terribly underwhelming, like +1 possible advisor, which is effectively useless for AI, so gave a modest buff in such cases. Tried to make things better fit the flavor; for example, regulating maritime trade now grants a merchant instead of a diplomat like in Vanilla.